// react18-app-admin/src/types/apis/sys/acl/aclType.ts
export interface AclModuleAddReq {
  name: string
  parentSurrogateId: string
  menuUrl?: string
  seq: number
  status: number
  remark: string
  operateIp?: string
}

export interface AclModuleEditReq extends AclModuleAddReq {
  surrogateId: string
}